separate some handlers to a library
