+++ app/src/main/res/layout/activity_handle_locale.xml
@@ -5,7 +5,7 @@
     android:id="@+id/activity_handle_locale"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.kazucocoa.droidtesthelper.HandleLocaleActivity">
+    tools:context=".HandleLocaleActivity">
 
     <TextView
         android:id="@+id/description_for_change_locale"
