+++ droidtesthelperlib/src/main/AndroidManifest.xml
@@ -1,0 +1,12 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.kazucocoa.droidtesthelperlib">
+
+    <uses-permission
+        android:name="android.permission.MANAGE_ACCOUNTS"
+        android:maxSdkVersion="22" />
+
+    <application android:label="@string/app_name" android:supportsRtl="true">
+
+    </application>
+
+</manifest>
